+++ test/tests/services/resources-getter.js
@@ -29,6 +29,7 @@ describe('Service > ResourcesGetter', () => {
             { field: 'comment', type: 'String' },
             { field: 'giftMessage', type: 'String' },
             { field: 'orderer', type: 'ObjectId', reference: 'User._id' },
+            { field: 'receiver', type: 'ObjectId', reference: 'User._id' },
           ],
         },
         User: {
@@ -53,6 +54,7 @@ describe('Service > ResourcesGetter', () => {
           comment: { type: String },
           giftMessage: { type: String },
           orderer: { type: 'ObjectId' },
+          receiver: { type: 'ObjectId' },
         });
         const UserSchema = mongoose.Schema({
           _id: { type: 'ObjectId' },
@@ -63,31 +65,37 @@ describe('Service > ResourcesGetter', () => {
         OrderModel = mongoose.model('Order', OrderSchema);
         UserModel = mongoose.model('User', UserSchema);
 
-        return OrderModel.remove({});
+        return Promise.all([OrderModel.remove({}), UserModel.remove({})]);
       })
       .then(() => {
-        loadFixture(OrderModel, [
-          {
-            // _id: 100,
-            amount: 199,
-            comment: 'no comment!',
-            giftMessage: 'Here is your gift',
-          },
-          {
-            // _id: 101,
-            amount: 1399,
-            comment: 'this is a gift',
-            giftMessage: 'Thank you',
-            orderer: '41224d776a326fb40f000001',
-          },
-        ]);
-
-        loadFixture(UserModel, [
-          {
-            _id: '41224d776a326fb40f000001',
-            age: 49,
-            name: 'Rust Cohle',
-          },
+        return Promise.all([
+          loadFixture(OrderModel, [
+            {
+              // _id: 100,
+              amount: 199,
+              comment: 'no comment!',
+              giftMessage: 'Here is your gift',
+              receiver: '41224d776a326fb40f000002',
+            },
+            {
+              // _id: 101,
+              amount: 1399,
+              comment: 'this is a gift',
+              giftMessage: 'Thank you',
+              orderer: '41224d776a326fb40f000001',
+            },
+          ]), loadFixture(UserModel, [
+            {
+              _id: '41224d776a326fb40f000001',
+              age: 49,
+              name: 'Rust Cohle',
+            },
+            {
+              _id: '41224d776a326fb40f000002',
+              age: 30,
+              name: 'Jacco Gardner',
+            },
+          ]),
         ]);
       });
   });
@@ -141,8 +149,7 @@ describe('Service > ResourcesGetter', () => {
           order: '_id,amount,description,giftMessage',
         },
         page: { number: '1', size: '30' },
-        filter: { giftMessage: 'Here*' },
-        filterType: 'and',
+        filters: JSON.stringify({ field: 'giftMessage', operator: 'starts_with', value: 'Here' }),
         timezone: '+02:00',
       };
 
@@ -164,8 +171,13 @@ describe('Service > ResourcesGetter', () => {
           order: '_id,amount,description,giftMessage',
         },
         page: { number: '1', size: '30' },
-        filter: { giftMessage: '*you*', amount: '>1000' },
-        filterType: 'and',
+        filters: JSON.stringify({
+          aggregator: 'and',
+          conditions: [
+            { field: 'giftMessage', operator: 'contains', value: 'you' },
+            { field: 'amount', operator: 'greater_than', value: '1000' },
+          ],
+        }),
         timezone: '+02:00',
       };
 
@@ -180,15 +192,14 @@ describe('Service > ResourcesGetter', () => {
     });
 
     describe('with belongsTo filter', () => {
-      describe('works with flat condition', () => {
+      describe('with flat condition', () => {
         it('should filter correctly', (done) => {
           const parameters = {
             fields: {
               order: '_id,amount,description,giftMessage',
             },
             page: { number: '1', size: '30' },
-            filterType: 'and',
-            filter: { 'orderer:name': '*Cohle*' },
+            filters: JSON.stringify({ field: 'orderer:name', operator: 'contains', value: 'Cohle' }),
             timezone: '+02:00',
           };
 
@@ -197,6 +208,100 @@ describe('Service > ResourcesGetter', () => {
             .then((result) => {
               expect(result[0].length).equal(1);
               expect(result[0][0].comment).to.match(/gift/);
+              done();
+            })
+            .catch(done);
+        });
+      });
+
+      describe('with \'and\' aggregator on two belongsTo on the same model', () => {
+        it('should filter correctly', (done) => {
+          const parameters = {
+            fields: {
+              order: '_id,amount,description,giftMessage',
+            },
+            page: { number: '1', size: '30' },
+            filters: JSON.stringify({
+              aggregator: 'and',
+              conditions: [
+                { field: 'orderer:name', operator: 'contains', value: 'Cohle' },
+                { field: 'orderer:name', operator: 'ends_with', value: 'Gardner' },
+              ],
+            }),
+            timezone: '+02:00',
+          };
+
+          new ResourcesGetter(OrderModel, options, parameters)
+            .perform()
+            .then((result) => {
+              expect(result[0].length).equal(0);
+              done();
+            })
+            .catch(done);
+        });
+      });
+
+      describe('with \'or\' aggregator on two belongsTo on the same model', () => {
+        it('should filter correctly', (done) => {
+          const parameters = {
+            fields: {
+              order: '_id,amount,description,giftMessage',
+            },
+            page: { number: '1', size: '30' },
+            filters: JSON.stringify({
+              aggregator: 'or',
+              conditions: [
+                { field: 'orderer:name', operator: 'contains', value: 'Cohle' },
+                { field: 'receiver:name', operator: 'ends_with', value: 'Gardner' },
+              ],
+            }),
+            timezone: '+02:00',
+          };
+
+          new ResourcesGetter(OrderModel, options, parameters)
+            .perform()
+            .then((result) => {
+              expect(result[0].length).equal(2);
+              done();
+            })
+            .catch(done);
+        });
+      });
+
+      describe('with complex nested filters', () => {
+        it('should filter correctly', (done) => {
+          const parameters = {
+            fields: {
+              order: '_id,amount,description,giftMessage',
+            },
+            page: { number: '1', size: '30' },
+            filters: JSON.stringify({
+              aggregator: 'or',
+              conditions: [
+                { field: 'orderer:name', operator: 'contains', value: 'elsewhere' },
+                {
+                  aggregator: 'and',
+                  conditions: [
+                    { field: 'giftMessage', operator: 'contains', value: 'you' },
+                    {
+                      aggregator: 'and',
+                      conditions: [
+                        { field: 'orderer:age', operator: 'blank', value: null },
+                        { field: 'receiver:name', operator: 'ends_with', value: 'Gardner' },
+                      ],
+                    },
+                    { field: 'amount', operator: 'less_than', value: '200' },
+                  ],
+                },
+              ],
+            }),
+            timezone: '+02:00',
+          };
+
+          new ResourcesGetter(OrderModel, options, parameters)
+            .perform()
+            .then((result) => {
+              expect(result[0].length).equal(1);
               done();
             })
             .catch(done);